feat: migrate discord page and fixed mobile issue
