+++ app/discord/page.tsx
@@ -1,0 +1,5 @@
+import { Discord } from './Discord';
+
+export default function Page() {
+	return <Discord />;
+}
